[refactor]:propaganda-route
